tongxinyuan-community/apps/ai-agent/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
asyncpg==0.29.0
python-dotenv==1.0.1
redis==5.0.1
httpx==0.26.0
pgvector==0.2.4
openai==1.10.0
dashscope==1.14.0 # Alibaba Tongyi Qianwen SDK
